// src/app/modules/voicebot/lesson/models/sentenceExampleExerciseModel.ts
export default class SentenceExampleExercisesModel {
  example: string | undefined = undefined;

  translatedExample: string | undefined = undefined;

  imageSrc: string | null = null;

  isImage: boolean | undefined = undefined;

  typeOfExample: string | undefined = undefined;

  something: string | null = null;
}